feat: add fileCollectionFromFileArray
closes: https://github.com/cheminfo/filelist-utils/issues/60
