+++ src/index.ts
@@ -3,6 +3,7 @@ export * from './fileCollectionFromPaths';
 export * from './fileCollectionFromZip';
 export * from './fileCollectionFromWebservice';
 export * from './fileCollectionFromFileList';
+export * from './fileCollectionFromFileArray';
 export * from './fileCollectionFromFiles';
 export * from './groupFiles';
 export * from './FileCollection';
